refactor(mcp): extract tasks domain
